+++ LinearAlgebra/tests/SVDecompTests.h
@@ -5,7 +5,7 @@
 using namespace LMFAO::LinearAlgebra;
 using namespace std;
 
-constexpr long double SVD_TEST_PRECISION_ERROR = 1e-10;
+constexpr double SVD_TEST_PRECISION_ERROR = 1e-10;
 
 TEST(SVDNaive, 2SizeCntMatrix) {
     static const string FILE_INPUT = "tests/data/test1/test.in";
